Tile and onion skin caching

+++ app/src/main/java/com/zarbosoft/pyxyzygy/app/parts/editor/OnionSkin.java
@@ -5,13 +5,17 @@ import com.zarbosoft.pyxyzygy.app.*;
 import com.zarbosoft.pyxyzygy.app.model.v0.ProjectContext;
 import com.zarbosoft.pyxyzygy.app.widgets.HelperJFX;
 import com.zarbosoft.pyxyzygy.core.TrueColorImage;
-import com.zarbosoft.pyxyzygy.core.mynative;
 import com.zarbosoft.pyxyzygy.nearestneighborimageview.NearestNeighborImageView;
+import com.zarbosoft.pyxyzygy.seed.model.v0.Rectangle;
 import com.zarbosoft.pyxyzygy.seed.model.v0.TrueColor;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+
+import static com.zarbosoft.pyxyzygy.app.GUILaunch.CACHE_ONION;
+import static com.zarbosoft.rendaw.common.Common.uncheck;
 
 public class OnionSkin {
 	final ImageView widget = NearestNeighborImageView.create();
@@ -69,15 +73,35 @@ public class OnionSkin {
 				.point(bounds.get().corner().minus(span.multiply(0.5)))
 				.point(bounds.get().corner().plus(span.multiply(1.5)))
 				.build();
-		final DoubleRectangle renderBounds = new BoundsBuilder()
+		final Rectangle renderBounds = new BoundsBuilder()
 				.point(bounds.get().corner().minus(span))
 				.point(bounds.get().corner().plus(span.multiply(2)))
-				.build();
-		final TrueColorImage buffer =
-				TrueColorImage.create((int) renderBounds.width + 1, (int) renderBounds.height + 1);
-		Render.render(context, editHandle.getWrapper().getValue(), buffer, frame, renderBounds.divideContains(1), 1);
-		widget.setImage(HelperJFX.toImage(buffer, color));
-		buffer.delete();
+				.build()
+				.divideContains(1);
+		if (renderBounds.height == 0 || renderBounds.width == 0)
+			return;
+		int hash = Objects.hashCode(CACHE_ONION,
+				editHandle.getWrapper().getValue().id(),
+				frame,
+				color.hashCode(),
+				renderBounds.hashCode()
+		);
+		Image image = uncheck(() -> GUILaunch.imageCache.get(hash, () -> {
+			final TrueColorImage buffer = TrueColorImage.create(renderBounds.width, renderBounds.height);
+			try {
+				Render.render(context,
+						editHandle.getWrapper().getValue(),
+						buffer,
+						frame,
+						renderBounds.divideContains(1),
+						1
+				);
+				return HelperJFX.toImage(buffer, color);
+			} finally {
+				buffer.delete();
+			}
+		}));
+		widget.setImage(image);
 		widget.setOpacity(color.toJfx().getOpacity());
 		widget.setX(renderBounds.x);
 		widget.setY(renderBounds.y);
